refactor: move CLI headless execution to runtime
